Add imagelibrary.json file generation, bug fixes
Thumbnails were being generated but not mapped to
the original image, imagelibrary.json now maps the
recently created thumbnails to the original images
also adding potential for more data in the future.

Corrected a bug in which default values of the
generateThumbnails command would not allow for
custom thumbnail sizes.

Updated README and general function documentation.