Add list deletion code and start dev on import
New feature in progress, import images, which will
copy the images from the given folder into the
gallery/images folder, from which we can continue
the generation script.

Small add-on to the list generation function:
now we can specify if we want to generate absolute
paths or relative paths, this is at the moment for
internal use only, if the necessity arises we can
add this as a command flag for the utility.

General style corrections