Add copy images to gallery function, dir checks
Before we were only listing and creating images
based on the list created, and depended on the
user to copy manually de images to the correct
folder.

Now we can specify a folder and then filter only
image files, and copy them to the correct folder
for further processing

The function name was changed from import to copy
as the action is to copy and map the images only
they are not "imported" per se?

Now the thumbnail creation function also checks
for the thumbs directory creation, and if it does
not exists, it will create it.